Rename JSdoc `component` parsing file
